Refactor & update to dotnet 8
