+++ .github/workflows/github-test-build.yml
@@ -8,7 +8,7 @@ on:
     branches: [ "main" ]
     paths:
       - Messaging/**
-      - SonicWallInterface/**
+      - FirewallBlocker/**
       - .github/workflows/github-test-build.yml
 
 jobs:
