+++ FirewallBlocker.Tests/FirewallBlocker.Tests.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
@@ -29,7 +29,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Messaging\Messaging.csproj" />
-    <ProjectReference Include="..\SonicWallInterface\SonicWallInterface.csproj" />
+    <ProjectReference Include="..\FirewallBlocker\FirewallBlocker.csproj" />
   </ItemGroup>
 
 </Project>
