+++ FirewallBlocker/Configuration/ServiceBusConfig.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 
-namespace SonicWallInterface.Configuration
+namespace FirewallBlocker.Configuration
 {
     public class ServiceBusConfig : IIsPresent
     {
