+++ FirewallBlocker/FirewallBlocker.csproj
@@ -2,11 +2,11 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <UserSecretsId>9a29c872-302c-4fb3-baea-c9b01650ed6e</UserSecretsId>
-    <Version>1.0.1</Version>
+    <Version>1.1.0</Version>
     <RuntimeIdentifiers>win-x64;linux-x64</RuntimeIdentifiers>
   </PropertyGroup>
