+++ FirewallBlocker/Services/HttpIPListApi.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace SonicWallInterface.Services
+namespace FirewallBlocker.Services
 {
     public class HttpIPListApi : IHttpIPListApi
     {
