+++ FirewallBlocker/Services/IFireWallApi.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SonicWallInterface.Services
+namespace FirewallBlocker.Services
 {
     public interface IFireWallApi
     {
