+++ FirewallBlocker/Services/ITIHandler.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace SonicWallInterface.Services
+namespace FirewallBlocker.Services
 {
     public interface ITIHandler
     {
